feat: included loading component at ticket selection while data is retrieved from the server / refac: adjusted ticket selection structuring
