+++ src/components/Splash.js
@@ -5,9 +5,9 @@ import Page from "./Page";
 
 import driventLogo from "../assets/images/drivent.png";
 
-export default function Splash({ loading = false, message = "" }) {
+export default function Splash({ loading = false, message = "", minHeight = "" }) {
   return (
-    <StyledPage background="#FA4098">
+    <StyledPage background="#FA4098" minHeight={minHeight}>
       <Grid>
         {loading && <StyledLoader color="#FFFFFF" height={26} width={26} type="Oval" />}
         <img src={driventLogo} alt="Driven.t" />
@@ -19,6 +19,7 @@ export default function Splash({ loading = false, message = "" }) {
 
 const StyledPage = styled(Page)`
   color: white;
+  min-height: ${({ minHeight }) => minHeight };
   padding: 20px;
 `;
